Revise HTTP API module structure
