+++ application/comit_node/src/http_api/routes/index/handlers/mod.rs
@@ -1,0 +1,3 @@
+mod get_swaps;
+
+pub use self::get_swaps::handle_get_swaps;
