+++ application/comit_node/src/http_api/routes/rfc003/handlers/mod.rs
@@ -1,13 +1,11 @@
 mod get_action;
 mod get_swap;
-mod get_swaps;
 mod post_action;
 mod post_swap;
 
 pub use self::{
     get_action::{handle_get_action, GetActionQueryParams},
     get_swap::handle_get_swap,
-    get_swaps::handle_get_swaps,
     post_action::handle_post_action,
-    post_swap::{handle_post_swap, SwapRequestBodyKind},
+    post_swap::{handle_post_swap, OnlyRedeem, OnlyRefund, SwapRequestBodyKind},
 };
